+++ src/routes/paths.ts
@@ -4,9 +4,12 @@ export const ROUTE_PATH = {
   OWNER: '/owner',
   LANDING: '/landing',
   LOGIN: '/login',
-  MY_PAGE: '/my',
-  MY_COUPON: '/my/coupon',
-  MY_LETTER: '/my/letter',
+  MY: '/my',
+  MY_CUSTOMER: '/my/customer',
+  MY_COUPON: '/my/customer/coupon',
+  MY_LETTER: '/my/customer/letter',
+  MY_OWNER: '/my/owner',
+  MY_MARKET: '/my/owner/market',
 
   MAP: '/map',
   SEARCH: '/search',
