+++ src/routes/router.tsx
@@ -8,11 +8,14 @@ import SearchPage from '@/pages/search/SearchPage';
 import StoreDetailPage from '@/pages/store/StoreDetailPage';
 import LetterPage from '@/pages/letter/LetterPage';
 import MyPage from '@/pages/my/MyPage';
-import MyCouponPage from '@/pages/my/coupon/MyCouponPage';
-import MyLetterPage from '@/pages/my/letter/MyLetterPage';
 import CouponPage from '@/pages/coupon/CouponPage';
 import CustomerMainPage from '@/pages/main/customer/CustomerMainPage';
 import OwnerMainPage from '@/pages/main/owner/OwnerMainPage';
+import CustomerMyPage from '@/pages/my/customer/CustomerMyPage';
+import OwnerMyPage from '@/pages/my/owner/OwnerMyPage';
+import MyMarketPage from '@/pages/my/owner/market/MyMarketPage';
+import MyCouponPage from '@/pages/my/customer/coupon/MyCouponPage';
+import MyLetterPage from '@/pages/my/customer/letter/MyLetterPage';
 
 const Router = () => {
   return (
@@ -28,9 +31,14 @@ const Router = () => {
       <Route path={ROUTE_PATH.STORE_DETAIL} element={<StoreDetailPage />} />
       <Route path={ROUTE_PATH.LETTER} element={<LetterPage />} />
       <Route path={ROUTE_PATH.COUPON} element={<CouponPage />} />
-      <Route path={ROUTE_PATH.MY_PAGE} element={<MyPage />}>
-        <Route path={ROUTE_PATH.MY_COUPON} element={<MyCouponPage />} />
-        <Route path={ROUTE_PATH.MY_LETTER} element={<MyLetterPage />} />
+      <Route path={ROUTE_PATH.MY} element={<MyPage />}>
+        <Route path={ROUTE_PATH.MY_CUSTOMER} element={<CustomerMyPage />}>
+          <Route path={ROUTE_PATH.MY_COUPON} element={<MyCouponPage />} />
+          <Route path={ROUTE_PATH.MY_LETTER} element={<MyLetterPage />} />
+        </Route>
+        <Route path={ROUTE_PATH.MY_OWNER} element={<OwnerMyPage />}>
+          <Route path={ROUTE_PATH.MY_MARKET} element={<MyMarketPage />} />
+        </Route>
       </Route>
     </Routes>
   );
